Update hooks directory to use hard links
The hooks directory should be linked to the existing hooks
repository. However, if soft links are used then the path of
each collaborators hooks directory will need to be the same. To
get around this hard links will be used. This has the added
benefit that we can see if hooks are up to date.